controller: Fix controller update
Don't stop the old goroutine to start a new one on update, to
prevent having two goroutines running simultaneously for a single
controller.
Instead, atomically change the parameters of the existing
controller.
Make sure that all fields (esp. lastError) are updated while
locking the controller mutex.

Preserving the old controller when it is updated also preserves
the controller's UUID and stats (success and failure counts, success
and failure timestamps, etc.).

Signed-off-by: Romain Lenglet <romain@covalent.io>